Extract dataInterfaces in separate file Add simple filter
